backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
bcrypt==4.1.2
Pillow>=11.0.0,<12.0.0
email-validator==2.1.0
waitress==2.1.2
gunicorn==21.2.0
